+++ packages/email/src/email.module.ts
@@ -1,8 +1,13 @@
 import { Module } from '@nestjs/common';
+import { BullModule } from '@nestjs/bullmq';
+import { EmailProcessor } from "@email/email.processor"
 import { EmailService } from './email.service';
 
 @Module({
-  providers: [EmailService],
+  imports: [
+    BullModule.registerQueue({ name: 'email' }),
+  ],
+  providers: [EmailService, EmailProcessor],
   exports: [EmailService],
 })
 export class EmailModule {}